feat: use dpr srcset for images with mark param
This commit refactors some of the `<img>` tag paprsing logic to allow
for using DPR srcset on images that use the `mark=` param. Before this
commit, there was an issue where the mark images had a different minimum
width than the `src` image. This could lead to issues when using the
mark as a layer to for example change the color of a product.